Fix line number anchors in gitweb links.
In one case, provide a commit hash so that the line number is relative
to a known commit of the file. I found the commit in bridgedb.git where
the URL was added, then found a commit in torspec.git that was around
the same time and seems to point to the right place.

Subtract one from the line numbers; maybe the old gitweb and the new are
off by one relative to each other.